+++ libs/core/src/thds/core/log/__init__.py
@@ -27,4 +27,11 @@ logger.info("testing 5")
 from . import env  # noqa: F401
 from .basic_config import DuplicateFilter, set_logger_to_console_level  # noqa: F401
 from .kw_formatter import ThdsCompactFormatter  # noqa: F401
-from .kw_logger import KwLogger, auto, getLogger, logger_context, make_th_formatters_safe  # noqa: F401
+from .kw_logger import (  # noqa: F401
+    KwLogger,
+    auto,
+    getLogger,
+    logger_context,
+    logger_context_values,
+    make_th_formatters_safe,
+)